+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## (Unreleased) - BREAKING
 
 - Add badges to the README
-- Add deny.toml, remove breakpad support to avoid MPL-2.0 dependencies (disallowed by bevy)
+- BREAKING: Add deny.toml, remove breakpad support to avoid MPL-2.0 dependencies (disallowed by bevy)
 
 ## 0.1.0 (2022-05-06)